Tidy the example
